+++ CHANGELOG.md
@@ -5,8 +5,10 @@
 * jre: published neowu/jre:26.0.1
 * stats: escalate to ERROR is cpu or memory high usage last more than 1 mins
   > call vm() and heap() every 1 min if memory usage is high, make it easier for AI troubleshoot
-* web: updated undertow to 2.4.1.Final
 * kafka: updated client to 4.3.0
+* web: updated undertow to 2.4.1.Final
+* web: replace path_not_found and forbidden warning with 404/403 and empty content
+  > reduce noisy for scanning requests, action log provides enough context
 
 ### 9.5.2 (4/29/26 - 5/4/26)